% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncells.R
\name{ncells}
\alias{ncells}
\title{Estimate the probability of success separation in two population model}
\usage{
ncells(p1, n1, foldchange, dropout, p = 26616, n = seq(100, 1000, by = 100),
  mu = NULL, tau = 5, perm = TRUE, dfactor = 2, seed = 123, B = 200,
  m = 50, progress = TRUE, ncore = 1)
}
\arguments{
\item{p1}{Fraction of marker genes among total number genes in the novel subpopulation, 0.1 <= \code{p1} <= 1.6}

\item{n1}{Fraction of cells in the novel subpopulation, 5 <= \code{n1} <= 10}

\item{foldchange}{Expression fold change in the novel subpopulation, 1.5 <= \code{foldchange} <= 8}

\item{dropout}{Overall dropout rate, 0.6 <= \code{dropout} <= 0.95}

\item{p}{Total number of genes in the simulation, 10000 <= \code{p}}

\item{n}{Vector of total number of cells, 100 <= min(n), max(n) <= 1000}

\item{mu}{Vector of mean expressions for \code{p} genes,}

\item{tau}{Standard deviation of mean expressions for \code{p} genes, 0 <= \code{tau}}

\item{perm}{Permute \code{mu}?}

\item{dfactor}{Multiplicative factor for separation criterion of two populations using first principal component scores, default is 2.}

\item{seed}{Random seed,}

\item{B}{Number of replicates to estiamte the probability of success separation, default is 200, 100 <= \code{B}}

\item{m}{Number of replicates to generated marginal distribution of expressions given \code{mu}, default is 50.}

\item{progress}{Print progress of simulations?}

\item{ncore}{Number of cores for use in the simulation using \code{\link{parallel}} package. default is 1, 1 <= \code{ncore} <= \code{parallel::detectCores()}}
}
\value{
Vector of estiamted probabilities of same length as \code{n}
}
\examples{
\dontrun{
## if mu is not specified, 
ncells(p1=1.6, n1=5, foldchange=4, dropout=.6)

## if mu is specified (for example as below)
p <- 26616
tau <- 5
mu <- rnorm(p)*tau
ncells(p1=1.6, n1=5, foldchange=4, dropout=.6, mu=mu)
}
}
